requirements.txt
python-dotenv
langchain
langchain-openai
openai